anti-replay: use uint32_t for the interface, document endiannes
The counter is a uint32_t so update the interface to reflect that. As it
is returned in pin data and is expected to be little-endian when parsed
by the pin server, document this and add FIXMEs where the bytes currently
aren't swapped for big-endian builds.

+++ main/storage.h
@@ -47,7 +47,7 @@ bool storage_get_encrypted_blob(uint8_t* encrypted, size_t encrypted_len, size_t
 bool storage_decrement_counter(void);
 bool storage_restore_counter(void);
 uint8_t storage_get_counter(void);
-bool storage_get_replay_counter(uint8_t* replay_counter);
+bool storage_get_replay_counter(uint32_t* replay_counter);
 bool storage_erase_encrypted_blob(void);
 
 bool storage_set_key_flags(uint8_t flags);
